chore(coverage): Improve branch coverage accuracy
Add c8 ignore comments to exclude files from coverage

This commit adds `/* c8 ignore next */` comments to multiple source
files to fix a bug with the coverage report.
Without the coverage comment, the first line of files with comments or
imports get treated as un 'uncovered' branch.

This fixes that for the current source files